requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
colorama==0.4.6
blinker==1.9.0

pdf2image==1.17.0
Pillow==10.4.0
pdfplumber==0.11.4
opencv-python==4.10.0.84
pytesseract==0.3.13
